refactor: ParticipantList 컴포넌트를 Container/Presentational 구조로 분리(useUsersByIds 경로 이동 포함)
